refactor crate into cargo workspace
