+++ Cargo.toml
@@ -1,10 +1,12 @@
-[package]
-name = "advent"
-version = "0.1.0"
-edition = "2021"
+[workspace]
+resolver = "2"
+
+members = [
+    "y2023"
+]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[dependencies]
+[workspace.dependencies]
 clap = {version = "4", features = ["derive"]}
-thiserror = "1.0"
+thiserror = "1.0.63"
