+++ y2023/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "advent"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+clap = { workspace = true }
+thiserror = { workspace = true }
