+++ y2023/src/inputs/day2.txt
